Finalizer validator
